Add fix for TRAC ticket #279 where the last record of the day was not using all packet data. Therefore we either read the previous dayfile or next dayfile to get remaining packet data.
Also found a issue/issue with mean and standard deviation values where the last digits were changing
depending on the order of calculating. The solution was to use double values to do calculations
and this produced a consistent value for mean and standard deviation.

View ticket to more info on issue and testing did.