Updated code to process correctly the previous day and next day for rtmon dayfiles. The code there only processed previous day and next day for moc dayfiles. I added the argument src=moc  or src=rtmon to signal which previous day and next day dayfiles to use to correctly create last records keyword values.
When processing "in" dayfiles that are from moc, then use src=moc, to trigger use
of moc previous and next day's dayfiles.

When processing "in" dayfiles that are from rtmon dayfiles, then use src=rtmon, to trigger use
of rtmon previous and next day's dayfiles.

This executable is not being used for hsb dayfiles and is not setup for hsb dayfile processing.