Add test mode and stale lock file detection.
Add second argument $test, a string prepended to email Subject: field
and, if present, set test mode. If a test mode run detects a lock file,
an email is sent to notify of the condition, then exit.

Derive email To: field from two lists, the first for all runs, and
the second is not used for test runs. For test runs, the threshold
(first) argument should be set to -1 to guarantee "anomaly" detection.
Test mode deletes the flag file that it generates.