Add test mode.
Add second argument $test, a string prepended to email Subject: field
and, if present, set test mode.

erive email To: field from two lists, the first for all runs, and
the second is not used for test runs. For test runs, the threshold
(first) argument should be set to -1 to guarantee "anomaly" detection.
Test mode deletes the flag file that it generates.